// src/users/controller/users.controller.interface.ts
import { Request, Response, NextFunction } from 'express';

export interface IUserController {
  register: (req: Request, res: Response, next: NextFunction) => void;
  login: (req: Request, res: Response, next: NextFunction) => void;
  info: (req: Request, res: Response, next: NextFunction) => void;
}